backend/pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "*"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
pydantic = "^2.0"
python-dotenv = "*"
black = "^25.11.0"
isort = "^7.0.0"
flake8 = "^7.3.0"
package-name = "^0.1"
requests = "^2.32.5"
fastapi-users = {extras = ["sqlalchemy"], version = "^15.0.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-jwt = "^0.3.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/__pycache__/*", "*/migrations/*"]

